+++ examples/multiple test specs Example/TestLib/TestLib/UnitTests5/Test5.m
@@ -1,0 +1,12 @@
+@import XCTest;
+
+@interface Test5 : XCTestCase
+@end
+
+@implementation Test5
+
+- (void)testExample {
+    XCTAssertTrue(YES);
+}
+
+@end